+++ CHANGELOG.md
@@ -9,7 +9,9 @@ Aus Gründen der Übersichtlichkeit sind bei älteren Versionen die einzelnen Pa
 ## [Unreleased]
 ### Added
 - Hilfsmittelnummer.formatted()/.unformatted() hinzugefügt (med-Modul)
+- Text.replaceSonderzeichen()
 ### Fixed
+- Formattierung Hilfsmittelnummer
 - Anführungsstriche und Gedankenstriche werden bei der Text-Konvertierung in andere Zeichensätze ersetzt
 
 ## [6.5.2] - 2025-10-19